\name{gssReligion-package}
\alias{gssReligion-package}
\alias{gssReligion}
\docType{package}
\title{
~~ Modify Gss for Religion ~~
}
\description{
~~ This package will make changes to GSS data approriate for religion analysis ~~
}
\details{
\tabular{ll}{
Package: \tab gssReligion\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-07\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
~~ randywreed@gmail.com ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ religionModify(df,"new_rda_name.rda") ~~
}
